refactor: remove WorkflowsValidationUtil
