--- hedera-node/hedera-app/src/main/java/com/hedera/node/app/workflows/WorkflowsValidationUtil.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (C) 2023 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hedera.node.app.workflows;
-
-import static com.hedera.hapi.node.base.ResponseCodeEnum.DUPLICATE_TRANSACTION;
-import static com.hedera.hapi.node.base.ResponseCodeEnum.EMPTY_TRANSACTION_BODY;
-import static java.util.Objects.requireNonNull;
-
-import com.hedera.hapi.node.transaction.TransactionBody;
-import com.hedera.node.app.spi.workflows.PreCheckException;
-import com.hedera.node.app.state.HederaRecordCache;
-import edu.umd.cs.findbugs.annotations.NonNull;
-import javax.inject.Inject;
-import javax.inject.Singleton;
-
-// TODO: add documentation
-@Singleton
-public class WorkflowsValidationUtil { // TODO: better name?
-
-    private final HederaRecordCache hederaRecordCache;
-
-    @Inject
-    public WorkflowsValidationUtil(@NonNull final HederaRecordCache hederaRecordCache) {
-        this.hederaRecordCache = requireNonNull(hederaRecordCache);
-    }
-
-    public void checkDuplicates(final TransactionBody txBody) throws PreCheckException {
-        if (txBody == null) {
-            throw new PreCheckException(EMPTY_TRANSACTION_BODY);
-        }
-
-        final var foundTransactionRecord = hederaRecordCache.getRecord(txBody.transactionIDOrThrow());
-        if (foundTransactionRecord != null) {
-            throw new PreCheckException(DUPLICATE_TRANSACTION);
-        }
-    }
-}
