+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/workflows/handle/HandleWorkflow.java
@@ -17,12 +17,15 @@
 package com.hedera.node.app.workflows.handle;
 
 import static com.hedera.hapi.node.base.ResponseCodeEnum.SUCCESS;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.DUPLICATE_TRANSACTION;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.EMPTY_TRANSACTION_BODY;
 import static java.util.Collections.emptyList;
 import static java.util.Objects.requireNonNull;
 
 import com.hedera.hapi.node.base.Key;
 import com.hedera.hapi.node.base.ResponseCodeEnum;
 import com.hedera.node.app.records.BlockRecordManager;
+import com.hedera.hapi.node.transaction.TransactionBody;
 import com.hedera.node.app.service.token.ReadableAccountStore;
 import com.hedera.node.app.services.ServiceScopeLookup;
 import com.hedera.node.app.signature.ExpandedSignaturePair;
@@ -36,7 +39,6 @@ import com.hedera.node.app.spi.workflows.PreCheckException;
 import com.hedera.node.app.state.HederaRecordCache;
 import com.hedera.node.app.state.HederaState;
 import com.hedera.node.app.workflows.TransactionChecker;
-import com.hedera.node.app.workflows.WorkflowsValidationUtil;
 import com.hedera.node.app.workflows.dispatcher.ReadableStoreFactory;
 import com.hedera.node.app.workflows.dispatcher.TransactionDispatcher;
 import com.hedera.node.app.workflows.handle.record.RecordListBuilder;
@@ -82,10 +84,7 @@ public class HandleWorkflow {
   private final ServiceScopeLookup serviceScopeLookup;
   private final ConfigProvider configProvider;
   private final InstantSource instantSource;
-  private final WorkflowsValidationUtil workflowsValidationUtil;
-
-  private final HederaRecordCache recordCache;
-  private final TransactionChecker transactionChecker;
+  private final HederaRecordCache hederaRecordCache;
 
   @Inject
   public HandleWorkflow(
@@ -99,23 +98,18 @@ public class HandleWorkflow {
       @NonNull final ServiceScopeLookup serviceScopeLookup,
       @NonNull final ConfigProvider configProvider,
       @NonNull final InstantSource instantSource,
-      @NonNull final HederaRecordCache recordCache,
-      @NonNull final TransactionChecker transactionChecker,
-      @NonNull final WorkflowsValidationUtil workflowsValidationUtil) {
+      @NonNull final HederaRecordCache hederaRecordCache) {
     this.networkInfo = requireNonNull(networkInfo, "networkInfo must not be null");
     this.preHandleWorkflow = requireNonNull(preHandleWorkflow, "preHandleWorkflow must not be null");
     this.dispatcher = requireNonNull(dispatcher, "dispatcher must not be null");
-    this.blockRecordManager = requireNonNull(blockRecordManager, "recordManager must not be null");
+    this.blockRecordManager = requireNonNull(blockRecordManager, "blockRecordManager must not be null");
     this.signatureExpander = requireNonNull(signatureExpander, "signatureExpander must not be null");
     this.signatureVerifier = requireNonNull(signatureVerifier, "signatureVerifier must not be null");
     this.checker = requireNonNull(checker, "checker must not be null");
     this.serviceScopeLookup = requireNonNull(serviceScopeLookup, "serviceScopeLookup must not be null");
     this.configProvider = requireNonNull(configProvider, "configProvider must not be null");
     this.instantSource = requireNonNull(instantSource, "instantSource must not be null");
-    this.recordCache = requireNonNull(recordCache, "recordCache must not be null");
-    this.transactionChecker = requireNonNull(transactionChecker, "transactionChecker must not be null");
-    this.workflowsValidationUtil =
-        requireNonNull(workflowsValidationUtil, "workflowsValidationUtil must not be null");
+    this.hederaRecordCache = requireNonNull(hederaRecordCache, "hederaRecordCache must not be null");
   }
 
   /**
@@ -162,78 +156,78 @@ public class HandleWorkflow {
 
         PreHandleResult preHandleResult = null;
         try {
-            // Setup configuration
-            final var configuration = configProvider.getConfiguration();
-            final var hederaConfig = configuration.getConfigData(HederaConfig.class);
+          // Setup configuration
+          final var configuration = configProvider.getConfiguration();
+          final var hederaConfig = configuration.getConfigData(HederaConfig.class);
 
-            preHandleResult = getCurrentPreHandleResult(state, platformEvent, platformTxn, configuration);
-            final var transactionInfo = preHandleResult.txInfo();
-            final var txBody = transactionInfo.txBody();
-            recordBuilder
-                    .transaction(transactionInfo.transaction())
-                    .transactionBytes(transactionInfo.signedBytes())
-                    .transactionID(txBody.transactionID())
-                    .memo(txBody.memo());
+          preHandleResult = getCurrentPreHandleResult(state, platformEvent, platformTxn, configuration);
+          final var transactionInfo = preHandleResult.txInfo();
+          final var txBody = transactionInfo.txBody();
+          recordBuilder
+              .transaction(transactionInfo.transaction())
+              .transactionBytes(transactionInfo.signedBytes())
+              .transactionID(txBody.transactionID())
+              .memo(txBody.memo());
 
-            // If pre-handle was successful, we return the result. Otherwise, we charge the node or throw an exception.
-            switch (preHandleResult.status()) {
-                case SO_FAR_SO_GOOD -> {
-                    /* All good nothing to do */
-                }
-                case NODE_DUE_DILIGENCE_FAILURE -> createPenaltyPayment();
-                case UNKNOWN_FAILURE -> throw new IllegalStateException("Pre-handle failed with unknown failure");
-                default -> throw new PreCheckException(preHandleResult.responseCode());
+          // If pre-handle was successful, we return the result. Otherwise, we charge the node or throw an exception.
+          switch (preHandleResult.status()) {
+            case SO_FAR_SO_GOOD -> {
+              /* All good nothing to do */
             }
+            case NODE_DUE_DILIGENCE_FAILURE -> createPenaltyPayment();
+            case UNKNOWN_FAILURE -> throw new IllegalStateException("Pre-handle failed with unknown failure");
+            default -> throw new PreCheckException(preHandleResult.responseCode());
+          }
 
-            // Check transaction duplication
-          workflowsValidationUtil.checkDuplicates(preHandleResult.txInfo().txBody());
+          // Check transaction duplication
+          checkDuplicates(preHandleResult.txInfo().txBody());
 
-            // Check all signature verifications. This will also wait, if validation is still ongoing.
-            final var timeout = hederaConfig.workflowVerificationTimeoutMS();
-            final var maxMillis = instantSource.millis() + timeout;
-            final var payerKeyVerification =
-                preHandleResult.verificationResults().get(preHandleResult.payerKey());
-            if (payerKeyVerification.get(timeout, TimeUnit.MILLISECONDS).failed()) {
-                throw new HandleException(ResponseCodeEnum.INVALID_SIGNATURE);
+          // Check all signature verifications. This will also wait, if validation is still ongoing.
+          final var timeout = hederaConfig.workflowVerificationTimeoutMS();
+          final var maxMillis = instantSource.millis() + timeout;
+          final var payerKeyVerification =
+              preHandleResult.verificationResults().get(preHandleResult.payerKey());
+          if (payerKeyVerification.get(timeout, TimeUnit.MILLISECONDS).failed()) {
+            throw new HandleException(ResponseCodeEnum.INVALID_SIGNATURE);
+          }
+          for (final var key : preHandleResult.requiredKeys()) {
+            final var remainingMillis = maxMillis - instantSource.millis();
+            if (remainingMillis <= 0) {
+              throw new TimeoutException("Verification of signatures timed out");
             }
-            for (final var key : preHandleResult.requiredKeys()) {
-                final var remainingMillis = maxMillis - instantSource.millis();
-                if (remainingMillis <= 0) {
-                    throw new TimeoutException("Verification of signatures timed out");
-                }
-                final var verification = preHandleResult.verificationResults().get(key);
-                if (verification.get(remainingMillis, TimeUnit.MILLISECONDS).failed()) {
-                    throw new HandleException(ResponseCodeEnum.INVALID_SIGNATURE);
-                }
+            final var verification = preHandleResult.verificationResults().get(key);
+            if (verification.get(remainingMillis, TimeUnit.MILLISECONDS).failed()) {
+              throw new HandleException(ResponseCodeEnum.INVALID_SIGNATURE);
             }
+          }
 
-            // Setup context
-            final var stack = new SavepointStackImpl(state, configuration);
-            final var verifier = new BaseHandleContextVerifier(hederaConfig, preHandleResult.verificationResults());
-            final var context = new HandleContextImpl(
-                    txBody,
-                    preHandleResult.payer(),
-                    preHandleResult.payerKey(),
-                    TransactionCategory.USER,
-                    recordBuilder,
-                    stack,
-                    verifier,
-                    recordListBuilder,
-                    checker,
-                    dispatcher,
-                    serviceScopeLookup,
-                    blockRecordManager,
-                    recordCache);
+          // Setup context
+          final var stack = new SavepointStackImpl(state, configuration);
+          final var verifier = new BaseHandleContextVerifier(hederaConfig, preHandleResult.verificationResults());
+          final var context = new HandleContextImpl(
+              txBody,
+              preHandleResult.payer(),
+              preHandleResult.payerKey(),
+              TransactionCategory.USER,
+              recordBuilder,
+              stack,
+              verifier,
+              recordListBuilder,
+              checker,
+              dispatcher,
+              serviceScopeLookup,
+              blockRecordManager,
+              hederaRecordCache);
 
-            // Dispatch the transaction to the handler
-            dispatcher.dispatchHandle(context);
+          // Dispatch the transaction to the handler
+          dispatcher.dispatchHandle(context);
 
-            // TODO: Finalize transaction with the help of the token service
+          // TODO: Finalize transaction with the help of the token service
 
-            recordBuilder.status(SUCCESS);
+          recordBuilder.status(SUCCESS);
 
-            // commit state
-            stack.commit();
+          // commit state
+          stack.commit();
         } catch (final PreCheckException e) {
             recordFailedTransaction(e.responseCode(), recordBuilder, recordListBuilder);
         } catch (final HandleException e) {
@@ -259,57 +253,68 @@ public class HandleWorkflow {
         final var recordListResult = recordListBuilder.build();
 
         if (preHandleResult != null) {
-            // FUTURE: This needs to be replaced by a proper implementation, as can be found in PR
-            // https://github.com/hashgraph/hedera-services/pull/7473
-            recordCache.add(
-                    0, preHandleResult.payer(), recordListResult.mainRecord().record(), consensusNow);
+          // FUTURE: This needs to be replaced by a proper implementation, as can be found in PR
+          // https://github.com/hashgraph/hedera-services/pull/7473
+          hederaRecordCache.add(
+              0, preHandleResult.payer(), recordListResult.mainRecord().record(), consensusNow);
         } else {
-            throw new IllegalStateException("pre handle result was null!");
+          throw new IllegalStateException("pre handle result was null!");
         }
 
-        blockRecordManager.endUserTransaction(recordListResult.recordStream(), state);
+      blockRecordManager.endUserTransaction(recordListResult.recordStream(), state);
     }
 
-    private void recordFailedTransaction(
-            @NonNull final ResponseCodeEnum status,
-            @NonNull final SingleTransactionRecordBuilderImpl recordBuilder,
-            @NonNull final RecordListBuilder recordListBuilder) {
-        recordBuilder.status(status);
-        recordListBuilder.revertChildRecordBuilders(recordBuilder);
-        // TODO: Finalize failed transaction with the help of token-service and commit required state changes
+  private void checkDuplicates(final TransactionBody txBody) throws PreCheckException {
+    if (txBody == null) {
+      throw new PreCheckException(EMPTY_TRANSACTION_BODY);
     }
 
-    /*
-     * This method gets all the verification data for the current transaction. If pre-handle was previously ran
-     * successfully, we only add the missing keys. If it did not run or an error occurred, we run it again.
-     * If there is a due diligence error, this method will return a CryptoTransfer to charge the node along with
-     * its verification data.
-     */
-    @NonNull
-    private PreHandleResult getCurrentPreHandleResult(
-            @NonNull final HederaState state,
-            @NonNull final ConsensusEvent platformEvent,
-            @NonNull final ConsensusTransaction platformTxn,
-            @NonNull final VersionedConfiguration configuration)
-            throws PreCheckException {
-        final var metadata = platformTxn.getMetadata();
-        // We do not know how long transactions are kept in memory. Clearing metadata to avoid keeping it for too long.
-        platformTxn.setMetadata(null);
+    final var foundTransactionRecord = hederaRecordCache.getRecord(txBody.transactionIDOrThrow());
+    if (foundTransactionRecord != null) {
+      throw new PreCheckException(DUPLICATE_TRANSACTION);
+    }
+  }
 
-        // First check if pre-handle was run before (in which case metadata is a PreHandleResult)
-        if (preHandleStillValid(configuration, metadata)) {
-            final var preHandleResult = (PreHandleResult) metadata;
+  private void recordFailedTransaction(
+      @NonNull final ResponseCodeEnum status,
+      @NonNull final SingleTransactionRecordBuilderImpl recordBuilder,
+      @NonNull final RecordListBuilder recordListBuilder) {
+    recordBuilder.status(status);
+    recordListBuilder.revertChildRecordBuilders(recordBuilder);
+    // TODO: Finalize failed transaction with the help of token-service and commit required state changes
+  }
 
-            // In case of due diligence error, we prepare a CryptoTransfer to charge the node and return immediately.
-            if (preHandleResult.status() == Status.NODE_DUE_DILIGENCE_FAILURE) {
-                return createPenaltyPayment();
-            }
+  /*
+   * This method gets all the verification data for the current transaction. If pre-handle was previously ran
+   * successfully, we only add the missing keys. If it did not run or an error occurred, we run it again.
+   * If there is a due diligence error, this method will return a CryptoTransfer to charge the node along with
+   * its verification data.
+   */
+  @NonNull
+  private PreHandleResult getCurrentPreHandleResult(
+      @NonNull final HederaState state,
+      @NonNull final ConsensusEvent platformEvent,
+      @NonNull final ConsensusTransaction platformTxn,
+      @NonNull final VersionedConfiguration configuration)
+      throws PreCheckException {
+    final var metadata = platformTxn.getMetadata();
+    // We do not know how long transactions are kept in memory. Clearing metadata to avoid keeping it for too long.
+    platformTxn.setMetadata(null);
 
-            // If pre-handle was successful, we need to add signatures that were not known at the time of pre-handle.
-            if (preHandleResult.status() == Status.SO_FAR_SO_GOOD) {
-                return addMissingSignatures(state, preHandleResult, configuration);
-            }
-        }
+    // First check if pre-handle was run before (in which case metadata is a PreHandleResult)
+    if (preHandleStillValid(configuration, metadata)) {
+      final var preHandleResult = (PreHandleResult) metadata;
+
+      // In case of due diligence error, we prepare a CryptoTransfer to charge the node and return immediately.
+      if (preHandleResult.status() == Status.NODE_DUE_DILIGENCE_FAILURE) {
+        return createPenaltyPayment();
+      }
+
+      // If pre-handle was successful, we need to add signatures that were not known at the time of pre-handle.
+      if (preHandleResult.status() == Status.SO_FAR_SO_GOOD) {
+        return addMissingSignatures(state, preHandleResult, configuration);
+      }
+    }
 
         // If we reach this point, either pre-handle was not run or it failed but may succeed now.
         // Therefore, we simply rerun pre-handle.
